% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cdf}
\alias{cdf}
\title{Cumulative distribution function}
\usage{
cdf(
  x,
  Fx = NULL,
  type = c("discrete", "continuous"),
  dist = c("binom", "pois", "geom", "unif", "exp", "norm"),
  ...
)
}
\arguments{
\item{x}{a numeric vector with the observed values of the random variable}

\item{Fx}{if NULL (default), then Fx is computed using one of the implemented distributions (see description of the dist argument below); otherwise the user must pass the values of Fx}

\item{type}{type of random variable; either discrete (default) or continuous; this argument is only necessary when Fx is passed by the user}

\item{dist}{optional; name of the distribution (currently the binomial, poisson and geometric distributions for discrete random variables, or uniform, exponential and normal distribution for continuous random variables)}

\item{...}{further arguments passed to or from other methods}
}
\value{
the desired plot of the cumulative distribution function
}
\description{
This function plots the cumulative distribution function (c.d.f) of a random variable
}
\examples{
\donttest{

# binomial:
x <- 0:10
cdf(x, dist = "binom", size = 10, prob = 0.7)

# normal:
x <- rnorm(100, mean = 10, sd = 2)
cdf(x, dist = "norm", mean = 10, sd = 2)
}
}
