% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mega_sena.R
\name{download_megasena}
\alias{download_megasena}
\title{Resultados da mega-sena}
\usage{
download_megasena(resultado = c("dezenas", "ganhadores"))
}
\arguments{
\item{resultado}{tipo de resultado desejado (dezenas sorteadas ou número de número de ganhadores da sena, quina e quadra)}
}
\value{
tibble/data.frame com os resultados da mega-sena.
}
\description{
Resultados da mega-sena
}
