% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{Geometric2}
\alias{Geometric2}
\alias{dgeom2}
\alias{pgeom2}
\alias{qgeom2}
\alias{rgeom2}
\title{The Geometric Distribution (Version 2)}
\usage{
dgeom2(x, prob, log = FALSE)

pgeom2(q, prob, lower.tail = TRUE, log.p = FALSE)

qgeom2(p, prob, lower.tail = TRUE, log.p = FALSE)

rgeom2(n, prob)
}
\arguments{
\item{x, q}{vector of quantiles representing the total number of Bernoulli trials (failures + sucess).}

\item{prob}{probability of success in each trial. 0 < prob <= 1.}

\item{log, log.p}{logical; if TRUE, probabilities p are given as log(p).}

\item{lower.tail}{logical; if TRUE (default), probabilities are \eqn{P[X \le x]}; otherwise, \eqn{P[X > x]}.}

\item{p}{vector of probabilities.}

\item{n}{number of observations. If length(n) > 1, the length is taken to be the number required.}

\item{size}{total number of successful trials (failures + one success), or dispersion parameter (the shape parameter of the gamma mixing distribution). Must be strictly positive, need not be integer.}
}
\value{
dgeom2 gives the density, pgeom2 gives the distribution function, qgeom2 gives the quantile function, and rgeom2 generates random deviates.

Invalid prob will result in return value NaN, with a warning.

The length of the result is determined by n for rgeom2, and is the maximum of the lengths of the numerical arguments for the other functions.

The numerical arguments other than n are recycled to the length of the result. Only the first elements of the logical arguments are used.

rgeom2 returns a vector of type \code{\link[base:integer]{integer}} unless generated values exceed the maximum representable integer when \code{\link[base:double]{double}} values are returned since R version 4.0.0.
}
\description{
The Geometric Distribution (Version 2)
}
\details{
The geometric distribution with prob = p has density
\deqn{
f(x) = p(1-p)^{x-1},
}
for x = 1, 2, 3, ..., and \eqn{0 < p \leq 1}.

If an element of x is not integer, the result of dgeom2 is zero, with a warning.

The quantile is defined as the smallest value x such that  \eqn{F(x) \ge p}, where F is the distribution function.
}
\seealso{
\code{\link[stats:Geometric]{Geometric}} for the original implementation of the geometric distribution.

\code{\link[est031:NegBinomial2]{NegBinomial2}} for the negative binomial (Version 2) which generalizes the geometric distribution (Version 2).
}
\concept{Discrete}
\concept{Univariate}
\keyword{distribution}
