% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_apps.R
\name{run_app}
\alias{run_app}
\title{Function to run shiny apps}
\usage{
run_app(app = "tcl_prop")
}
\arguments{
\item{app}{name of the app}
}
\description{
This functions is used to run shiny apps implemented in the package.
}
