% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cdf}
\alias{cdf}
\title{Cumulative distribution function}
\usage{
cdf(
  x,
  Fx = NULL,
  dist = c("binom", "pois", "geom", "unif", "exp", "norm"),
  ...
)
}
\arguments{
\item{x}{a numeric vector with the observed values of the random variable.}

\item{Fx}{if NULL (default), then Fx is computed based one of the implemented distributions (currently the binomial, poisson and geometric distributions for discrete random variables, or uniform, exponential and normal distribution for continuous random variables); otherwise the user must pass the values of Fx.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
the desired plot of the cumulative distribution function.
}
\description{
This function plots the cumulative distribution function (c.d.f) of a random variable.
}
